+++ src/main/java/usyd/mingyi/animalcare/service/FriendService.java
@@ -6,12 +6,12 @@ import java.util.List;
 
 public interface FriendService {
 
-    int sendFriendRequest(int fromId,int toId);
+    int sendFriendRequest(long fromId,long toId);
 
-    int acceptFriendRequest(int fromId,int toId);
-    int rejectFriendRequest(int fromId,int toId);
-    List<User> getAllFriends(int id);//根据用户id获取用户所有的friends
-    List<User> getAllRequests(int id);//根据用户id获取用户所有的request
-    int checkFriendshipStatus(int fromId, int toId);
-    int deleteFromFriendList(int fromId,int toId);
+    int acceptFriendRequest(long fromId,long toId);
+    int rejectFriendRequest(long fromId,long toId);
+    List<User> getAllFriends(long id);//根据用户id获取用户所有的friends
+    List<User> getAllRequests(long id);//根据用户id获取用户所有的request
+    int checkFriendshipStatus(long fromId, long toId);
+    int deleteFromFriendList(long fromId,long toId);
 }
