+++ src/main/java/usyd/mingyi/animalcare/service/FriendServiceImp.java
@@ -15,7 +15,7 @@ public class FriendServiceImp implements FriendService {
 
     @Override
     @Transactional
-    public int sendFriendRequest(int fromId, int toId) {
+    public int sendFriendRequest(long fromId, long toId) {
         // 返回值 0失败 1成功 2之前已经收到过对方请求直接互为好友
         int i = friendMapper.checkRequestReverse(fromId, toId);
         if (i > 0) {
@@ -40,7 +40,7 @@ public class FriendServiceImp implements FriendService {
 
     @Override
     @Transactional
-    public int acceptFriendRequest(int fromId, int toId) {
+    public int acceptFriendRequest(long fromId, long toId) {
         if (friendMapper.checkExistRequest(toId, fromId) > 0 && !friendMapper.isFriend(fromId, toId)) {
             friendMapper.addToFriendList(fromId, toId);
             friendMapper.addToFriendList(toId, fromId);
@@ -55,23 +55,23 @@ public class FriendServiceImp implements FriendService {
     }
 
     @Override
-    public int rejectFriendRequest(int fromId, int toId) {
+    public int rejectFriendRequest(long fromId, long toId) {
         return friendMapper.deleteFriendRequest(toId, fromId);
     }
 
     @Override
-    public List<User> getAllFriends(int id) {
+    public List<User> getAllFriends(long id) {
         return friendMapper.getAllFriends(id);
     }
 
     @Override
-    public List<User> getAllRequests(int id) {
+    public List<User> getAllRequests(long id) {
         return friendMapper.getAllRequests(id);
     }
 
     @Override
     @Transactional
-    public int checkFriendshipStatus(int fromId, int toId) {
+    public int checkFriendshipStatus(long fromId, long toId) {
         // has already been friend
         if (friendMapper.isFriend(fromId, toId)) return 1;
         // has already sent friend request
@@ -82,7 +82,7 @@ public class FriendServiceImp implements FriendService {
 
     @Override
     @Transactional
-    public int deleteFromFriendList(int fromId, int toId) {
+    public int deleteFromFriendList(long fromId, long toId) {
         int delete = friendMapper.deleteFromFriendList(fromId, toId);
         int deleteReverse = friendMapper.deleteFromFriendList(toId, fromId);
         if (delete == 1 && deleteReverse == 1) {
